ml-services/requirements.txt
torch==2.0.0
torchvision==0.15.0
opencv-python==4.8.0
numpy==1.24.0
pillow==10.0.0
fastapi==0.100.0
uvicorn==0.23.0
python-multipart==0.0.6
pydantic==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.0
alembic==1.11.0
pytest==7.4.0
black==23.7.0
isort==5.12.0
flake8==6.1.0 